src/local_llm/requirements.txt
# Local LLM Stack - Python Dependencies
# Install with: pip install -r src/local_llm/requirements.txt

# Core CLI dependencies
typer>=0.9.0
rich>=13.0.0
pyyaml>=6.0

# API client
openai>=1.0.0
requests>=2.31.0

# Optional: Better performance
# psutil>=5.9.0  # Uncomment for enhanced system monitoring
